extract file related functions
