+++ src/lib.rs
@@ -3,3 +3,5 @@ pub use crate::page::genpage;
 
 mod app;
 pub use crate::app::parse_arg;
+
+pub mod files;
